Move tagged int ops to int_ops.c
