+++ mypyc/common.py
@@ -35,6 +35,7 @@ IS_32_BIT_PLATFORM = sys.maxsize < (1 << 31)  # type: Final
 RUNTIME_C_FILES = [
     'CPy.c',
     'getargs.c',
+    'int_ops.c',
     'list_ops.c',
     'dict_ops.c',
     'str_ops.c',
