+++ mypyc/lib-rt/CPy.h
@@ -152,11 +152,30 @@ static bool _CPy_IsSafeMetaClass(PyTypeObject *metaclass) {
     ((method_type)(CPy_FindTraitVtable(trait, ((object_type *)obj)->vtable)[vtable_index]))
 
 
-static void CPyError_OutOfMemory(void) {
-    fprintf(stderr, "fatal: out of memory\n");
-    fflush(stderr);
-    abort();
-}
+// Int operations
+
+CPyTagged CPyTagged_FromSsize_t(Py_ssize_t value);
+CPyTagged CPyTagged_FromObject(PyObject *object);
+CPyTagged CPyTagged_StealFromObject(PyObject *object);
+CPyTagged CPyTagged_BorrowFromObject(PyObject *object);
+PyObject *CPyTagged_AsObject(CPyTagged x);
+PyObject *CPyTagged_StealAsObject(CPyTagged x);
+Py_ssize_t CPyTagged_AsSsize_t(CPyTagged x);
+void CPyTagged_IncRef(CPyTagged x);
+void CPyTagged_DecRef(CPyTagged x);
+void CPyTagged_XDecRef(CPyTagged x);
+CPyTagged CPyTagged_Negate(CPyTagged num);
+CPyTagged CPyTagged_Add(CPyTagged left, CPyTagged right);
+CPyTagged CPyTagged_Subtract(CPyTagged left, CPyTagged right);
+CPyTagged CPyTagged_Multiply(CPyTagged left, CPyTagged right);
+CPyTagged CPyTagged_FloorDivide(CPyTagged left, CPyTagged right);
+CPyTagged CPyTagged_Remainder(CPyTagged left, CPyTagged right);
+bool CPyTagged_IsEq_(CPyTagged left, CPyTagged right);
+bool CPyTagged_IsLt_(CPyTagged left, CPyTagged right);
+PyObject *CPyTagged_Str(CPyTagged n);
+PyObject *CPyLong_FromStrWithBase(PyObject *o, CPyTagged base);
+PyObject *CPyLong_FromStr(PyObject *o);
+PyObject *CPyLong_FromFloat(PyObject *o);
 
 static inline int CPyTagged_CheckLong(CPyTagged x) {
     return x & CPY_INT_TAG;
@@ -182,146 +201,9 @@ static inline bool CPyTagged_TooBig(Py_ssize_t value) {
         && (value >= 0 || value < CPY_TAGGED_MIN);
 }
 
-static CPyTagged CPyTagged_FromSsize_t(Py_ssize_t value) {
-    // We use a Python object if the value shifted left by 1 is too
-    // large for Py_ssize_t
-    if (CPyTagged_TooBig(value)) {
-        PyObject *object = PyLong_FromSsize_t(value);
-        return ((CPyTagged)object) | CPY_INT_TAG;
-    } else {
-        return value << 1;
-    }
-}
-
-static CPyTagged CPyTagged_FromObject(PyObject *object) {
-    int overflow;
-    // The overflow check knows about CPyTagged's width
-    Py_ssize_t value = CPyLong_AsSsize_tAndOverflow(object, &overflow);
-    if (overflow != 0) {
-        Py_INCREF(object);
-        return ((CPyTagged)object) | CPY_INT_TAG;
-    } else {
-        return value << 1;
-    }
-}
-
-static CPyTagged CPyTagged_StealFromObject(PyObject *object) {
-    int overflow;
-    // The overflow check knows about CPyTagged's width
-    Py_ssize_t value = CPyLong_AsSsize_tAndOverflow(object, &overflow);
-    if (overflow != 0) {
-        return ((CPyTagged)object) | CPY_INT_TAG;
-    } else {
-        Py_DECREF(object);
-        return value << 1;
-    }
-}
-
-static CPyTagged CPyTagged_BorrowFromObject(PyObject *object) {
-    int overflow;
-    // The overflow check knows about CPyTagged's width
-    Py_ssize_t value = CPyLong_AsSsize_tAndOverflow(object, &overflow);
-    if (overflow != 0) {
-        return ((CPyTagged)object) | CPY_INT_TAG;
-    } else {
-        return value << 1;
-    }
-}
-
-static PyObject *CPyTagged_AsObject(CPyTagged x) {
-    PyObject *value;
-    if (CPyTagged_CheckLong(x)) {
-        value = CPyTagged_LongAsObject(x);
-        Py_INCREF(value);
-    } else {
-        value = PyLong_FromSsize_t(CPyTagged_ShortAsSsize_t(x));
-        if (value == NULL) {
-            CPyError_OutOfMemory();
-        }
-    }
-    return value;
-}
-
-static PyObject *CPyTagged_StealAsObject(CPyTagged x) {
-    PyObject *value;
-    if (CPyTagged_CheckLong(x)) {
-        value = CPyTagged_LongAsObject(x);
-    } else {
-        value = PyLong_FromSsize_t(CPyTagged_ShortAsSsize_t(x));
-        if (value == NULL) {
-            CPyError_OutOfMemory();
-        }
-    }
-    return value;
-}
-
-static Py_ssize_t CPyTagged_AsSsize_t(CPyTagged x) {
-    if (CPyTagged_CheckShort(x)) {
-        return CPyTagged_ShortAsSsize_t(x);
-    } else {
-        return PyLong_AsSsize_t(CPyTagged_LongAsObject(x));
-    }
-}
-
-CPy_NOINLINE
-static void CPyTagged_IncRef(CPyTagged x) {
-    if (CPyTagged_CheckLong(x)) {
-        Py_INCREF(CPyTagged_LongAsObject(x));
-    }
-}
-
-CPy_NOINLINE
-static void CPyTagged_DecRef(CPyTagged x) {
-    if (CPyTagged_CheckLong(x)) {
-        Py_DECREF(CPyTagged_LongAsObject(x));
-    }
-}
-
-CPy_NOINLINE
-static void CPyTagged_XDecRef(CPyTagged x) {
-    if (CPyTagged_CheckLong(x)) {
-        Py_XDECREF(CPyTagged_LongAsObject(x));
-    }
-}
-
 static inline bool CPyTagged_IsAddOverflow(CPyTagged sum, CPyTagged left, CPyTagged right) {
     // This check was copied from some of my old code I believe that it works :-)
     return (Py_ssize_t)(sum ^ left) < 0 && (Py_ssize_t)(sum ^ right) < 0;
-}
-
-static CPyTagged CPyTagged_Negate(CPyTagged num) {
-    if (CPyTagged_CheckShort(num)
-            && num != (CPyTagged) ((Py_ssize_t)1 << (CPY_INT_BITS - 1))) {
-        // The only possibility of an overflow error happening when negating a short is if we
-        // attempt to negate the most negative number.
-        return -num;
-    }
-    PyObject *num_obj = CPyTagged_AsObject(num);
-    PyObject *result = PyNumber_Negative(num_obj);
-    if (result == NULL) {
-        CPyError_OutOfMemory();
-    }
-    Py_DECREF(num_obj);
-    return CPyTagged_StealFromObject(result);
-}
-
-static CPyTagged CPyTagged_Add(CPyTagged left, CPyTagged right) {
-    // TODO: Use clang/gcc extension __builtin_saddll_overflow instead.
-    if (CPyTagged_CheckShort(left) && CPyTagged_CheckShort(right)) {
-        CPyTagged sum = left + right;
-        if (!CPyTagged_IsAddOverflow(sum, left, right)) {
-            return sum;
-        }
-    }
-    PyObject *left_obj = CPyTagged_AsObject(left);
-    PyObject *right_obj = CPyTagged_AsObject(right);
-    PyObject *result = PyNumber_Add(left_obj, right_obj);
-    if (result == NULL) {
-        CPyError_OutOfMemory();
-    }
-    Py_DECREF(left_obj);
-    Py_DECREF(right_obj);
-    return CPyTagged_StealFromObject(result);
 }
 
 static inline bool CPyTagged_IsSubtractOverflow(CPyTagged diff, CPyTagged left, CPyTagged right) {
@@ -329,75 +211,13 @@ static inline bool CPyTagged_IsSubtractOverflow(CPyTagged diff, CPyTagged left, 
     return (Py_ssize_t)(diff ^ left) < 0 && (Py_ssize_t)(diff ^ right) >= 0;
 }
 
-static CPyTagged CPyTagged_Subtract(CPyTagged left, CPyTagged right) {
-    // TODO: Use clang/gcc extension __builtin_saddll_overflow instead.
-    if (CPyTagged_CheckShort(left) && CPyTagged_CheckShort(right)) {
-        CPyTagged diff = left - right;
-        if (!CPyTagged_IsSubtractOverflow(diff, left, right)) {
-            return diff;
-        }
-    }
-    PyObject *left_obj = CPyTagged_AsObject(left);
-    PyObject *right_obj = CPyTagged_AsObject(right);
-    PyObject *result = PyNumber_Subtract(left_obj, right_obj);
-    if (result == NULL) {
-        CPyError_OutOfMemory();
-    }
-    Py_DECREF(left_obj);
-    Py_DECREF(right_obj);
-    return CPyTagged_StealFromObject(result);
-}
-
 static inline bool CPyTagged_IsMultiplyOverflow(CPyTagged left, CPyTagged right) {
     // This is conservative -- return false only in a small number of all non-overflow cases
     return left >= (1U << (CPY_INT_BITS/2 - 1)) || right >= (1U << (CPY_INT_BITS/2 - 1));
 }
 
-static CPyTagged CPyTagged_Multiply(CPyTagged left, CPyTagged right) {
-    // TODO: Consider using some clang/gcc extension
-    if (CPyTagged_CheckShort(left) && CPyTagged_CheckShort(right)) {
-        if (!CPyTagged_IsMultiplyOverflow(left, right)) {
-            return left * CPyTagged_ShortAsSsize_t(right);
-        }
-    }
-    PyObject *left_obj = CPyTagged_AsObject(left);
-    PyObject *right_obj = CPyTagged_AsObject(right);
-    PyObject *result = PyNumber_Multiply(left_obj, right_obj);
-    if (result == NULL) {
-        CPyError_OutOfMemory();
-    }
-    Py_DECREF(left_obj);
-    Py_DECREF(right_obj);
-    return CPyTagged_StealFromObject(result);
-}
-
 static inline bool CPyTagged_MaybeFloorDivideFault(CPyTagged left, CPyTagged right) {
     return right == 0 || left == -((size_t)1 << (CPY_INT_BITS-1));
-}
-
-static CPyTagged CPyTagged_FloorDivide(CPyTagged left, CPyTagged right) {
-    if (CPyTagged_CheckShort(left) && CPyTagged_CheckShort(right)
-        && !CPyTagged_MaybeFloorDivideFault(left, right)) {
-        Py_ssize_t result = ((Py_ssize_t)left / CPyTagged_ShortAsSsize_t(right)) & ~1;
-        if (((Py_ssize_t)left < 0) != (((Py_ssize_t)right) < 0)) {
-            if (result / 2 * right != left) {
-                // Round down
-                result -= 2;
-            }
-        }
-        return result;
-    }
-    PyObject *left_obj = CPyTagged_AsObject(left);
-    PyObject *right_obj = CPyTagged_AsObject(right);
-    PyObject *result = PyNumber_FloorDivide(left_obj, right_obj);
-    Py_DECREF(left_obj);
-    Py_DECREF(right_obj);
-    // Handle exceptions honestly because it could be ZeroDivisionError
-    if (result == NULL) {
-        return CPY_INT_TAG;
-    } else {
-        return CPyTagged_StealFromObject(result);
-    }
 }
 
 static inline bool CPyTagged_MaybeRemainderFault(CPyTagged left, CPyTagged right) {
@@ -406,41 +226,6 @@ static inline bool CPyTagged_MaybeRemainderFault(CPyTagged left, CPyTagged right
     // -1 is actually represented as -2 and can't overflow.
     // Mod by 0 can still fault though.
     return right == 0;
-}
-
-static CPyTagged CPyTagged_Remainder(CPyTagged left, CPyTagged right) {
-    if (CPyTagged_CheckShort(left) && CPyTagged_CheckShort(right)
-        && !CPyTagged_MaybeRemainderFault(left, right)) {
-        Py_ssize_t result = (Py_ssize_t)left % (Py_ssize_t)right;
-        if (((Py_ssize_t)right < 0) != ((Py_ssize_t)left < 0) && result != 0) {
-            result += right;
-        }
-        return result;
-    }
-    PyObject *left_obj = CPyTagged_AsObject(left);
-    PyObject *right_obj = CPyTagged_AsObject(right);
-    PyObject *result = PyNumber_Remainder(left_obj, right_obj);
-    Py_DECREF(left_obj);
-    Py_DECREF(right_obj);
-    // Handle exceptions honestly because it could be ZeroDivisionError
-    if (result == NULL) {
-        return CPY_INT_TAG;
-    } else {
-        return CPyTagged_StealFromObject(result);
-    }
-}
-
-static bool CPyTagged_IsEq_(CPyTagged left, CPyTagged right) {
-    if (CPyTagged_CheckShort(right)) {
-        return false;
-    } else {
-        int result = PyObject_RichCompareBool(CPyTagged_LongAsObject(left),
-                                              CPyTagged_LongAsObject(right), Py_EQ);
-        if (result == -1) {
-            CPyError_OutOfMemory();
-        }
-        return result;
-    }
 }
 
 static inline bool CPyTagged_IsEq(CPyTagged left, CPyTagged right) {
@@ -457,18 +242,6 @@ static inline bool CPyTagged_IsNe(CPyTagged left, CPyTagged right) {
     } else {
         return !CPyTagged_IsEq_(left, right);
     }
-}
-
-static bool CPyTagged_IsLt_(CPyTagged left, CPyTagged right) {
-    PyObject *left_obj = CPyTagged_AsObject(left);
-    PyObject *right_obj = CPyTagged_AsObject(right);
-    int result = PyObject_RichCompareBool(left_obj, right_obj, Py_LT);
-    Py_DECREF(left_obj);
-    Py_DECREF(right_obj);
-    if (result == -1) {
-        CPyError_OutOfMemory();
-    }
-    return result;
 }
 
 static inline bool CPyTagged_IsLt(CPyTagged left, CPyTagged right) {
@@ -503,60 +276,11 @@ static inline bool CPyTagged_IsLe(CPyTagged left, CPyTagged right) {
     }
 }
 
-static CPyTagged CPyTagged_Id(PyObject *o) {
-    return CPyTagged_FromSsize_t((Py_ssize_t)o);
-}
 
-
-#define MAX_INT_CHARS 22
-#define _PyUnicode_LENGTH(op)                           \
-    (((PyASCIIObject *)(op))->length)
-
-// using snprintf or PyUnicode_FromFormat was way slower than
-// boxing the int and calling PyObject_Str on it, so we implement our own
-static int fmt_ssize_t(char *out, Py_ssize_t n) {
-	bool neg = n < 0;
-	if (neg) n = -n;
-
-	// buf gets filled backward and then we copy it forward
-	char buf[MAX_INT_CHARS];
-	int i = 0;
-	do {
-		buf[i] = (n % 10) + '0';
-		n /= 10;
-		i++;
-	} while (n);
-
-
-	int len = i;
-	int j = 0;
-	if (neg) {
-		out[j++] = '-';
-		len++;
-	}
-
-	for (; j < len; j++, i--) {
-		out[j] = buf[i-1];
-	}
-	out[j] = '\0';
-
-	return len;
-}
-
-static PyObject *CPyTagged_ShortToStr(Py_ssize_t n) {
-	PyObject *obj = PyUnicode_New(MAX_INT_CHARS, 127);
-	if (!obj) return NULL;
-	int len = fmt_ssize_t((char *)PyUnicode_1BYTE_DATA(obj), n);
-	_PyUnicode_LENGTH(obj) = len;
-	return obj;
-}
-
-static PyObject *CPyTagged_Str(CPyTagged n) {
-	if (CPyTagged_CheckShort(n)) {
-		return CPyTagged_ShortToStr(CPyTagged_ShortAsSsize_t(n));
-	} else {
-		return PyObject_Str(CPyTagged_AsObject(n));
-	}
+static void CPyError_OutOfMemory(void) {
+    fprintf(stderr, "fatal: out of memory\n");
+    fflush(stderr);
+    abort();
 }
 
 static PyObject *CPyBool_Str(bool b) {
@@ -601,25 +325,6 @@ static inline int CPy_ObjectToStatus(PyObject *obj) {
 // might be an int also
 static inline bool CPyFloat_Check(PyObject *o) {
     return PyFloat_Check(o) || PyLong_Check(o);
-}
-
-static PyObject *CPyLong_FromFloat(PyObject *o) {
-    if (PyLong_Check(o)) {
-        CPy_INCREF(o);
-        return o;
-    } else {
-        return PyLong_FromDouble(PyFloat_AS_DOUBLE(o));
-    }
-}
-
-static PyObject *CPyLong_FromStrWithBase(PyObject *o, CPyTagged base) {
-        Py_ssize_t base_size_t = CPyTagged_AsSsize_t(base);
-        return PyLong_FromUnicodeObject(o, base_size_t);
-}
-
-static PyObject *CPyLong_FromStr(PyObject *o) {
-    CPyTagged base = CPyTagged_FromSsize_t(10);
-    return CPyLong_FromStrWithBase(o, base);
 }
 
 // Construct a nicely formatted type name based on __module__ and __name__.
@@ -913,6 +618,7 @@ int CPyDataclass_SleightOfHand(PyObject *dataclass_dec, PyObject *tp,
                                PyObject *dict, PyObject *annotations);
 PyObject *CPyPickle_SetState(PyObject *obj, PyObject *state);
 PyObject *CPyPickle_GetState(PyObject *obj);
+CPyTagged CPyTagged_Id(PyObject *o);
 
 
 // Check that dictionary didn't change size during iteration.
