+++ mypyc/lib-rt/setup.py
@@ -9,7 +9,7 @@ setup(name='test_capi',
       version='0.1',
       ext_modules=[Extension(
           'test_capi',
-          ['test_capi.cc', 'CPy.cc', 'list_ops.c'],
+          ['test_capi.cc', 'CPy.cc', 'int_ops.c', 'list_ops.c'],
           depends=['CPy.h', 'mypyc_util.h', 'pythonsupport.h'],
           extra_compile_args=['--std=c++11', '-Wno-unused-function', '-Wno-sign-compare'],
           library_dirs=['../external/googletest/make'],
